requirements.txt
albumentations==1.3.1
appdirs==1.4.4
brotlipy==0.7.0
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.3
cycler==0.11.0
cytoolz==0.12.0
dask==2023.10.1
docker-pycreds==0.4.0
filelock==3.9.0
fonttools==4.25.0
fsspec==2023.10.0
gitdb==4.0.10
gitpython==3.1.37
idna==3.4
imagecodecs==2021.8.26
imageio==2.31.5
importlib-metadata==6.8.0
jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
lazy-loader==0.3
locket==1.0.0
markupsafe==2.1.1
matplotlib==3.8.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
monai==1.3.0
mpmath==1.3.0
munkres==1.1.4
networkx==3.1
numpy==1.26.0
packaging==23.2
partd==1.4.1
pathtools==0.1.2
pillow==10.0.1
pip==23.2.1
ply==3.11
protobuf==4.24.4
psutil==5.9.6
pycparser==2.21
pyopenssl==23.2.0
pyparsing==3.0.9
pyqt5-sip==12.13.0
pyqt5==5.15.10
pysocks==1.7.1
python-dateutil==2.8.2
python-utils==3.3.3
pywavelets==1.4.1
pyyaml==6.0
qudida==0.0.4
requests==2.31.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.11.3
sentry-sdk==1.32.0
setproctitle==1.3.3
setuptools==68.0.0
sip==6.7.12
six==1.16.0
smmap==5.0.1
sympy==1.11.1
threadpoolctl==3.2.0
tifffile==2021.4.8
toolz==0.12.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.65.0
typing-extensions==4.7.1
urllib3==1.26.16
wandb==0.15.12
wheel==0.41.2
win-inet-pton==1.1.0
zipp==3.17.0